move more string stuff to config_strs
