+++ slk/create_config_files.py
@@ -18,21 +18,19 @@ The configs_dir (where the config files will reside) can be set through the comm
 or the environment variable RIPPLED_SIDECHAIN_CFG_DIR
 """
 
-import json
 import sys
 import traceback
 from pathlib import Path
-from typing import Dict, List, Optional, Tuple
+from typing import Dict, List, Optional
 
 from xrpl.models import IssuedCurrencyAmount
-from xrpl.wallet import Wallet
 
 from slk.chain import single_node_chain
 from slk.common import eprint
-from slk.config_classes import Keypair, Network, Ports, SidechainNetwork, XChainAsset
+from slk.config_classes import Network, Ports, SidechainNetwork, XChainAsset
 from slk.config_file import ConfigFile
 from slk.config_params import ConfigParams
-from slk.config_strs import get_cfg_str, get_ips_stanza
+from slk.config_strs import generate_sidechain_stanza, get_cfg_str, get_ips_stanza
 
 MAINNET_VALIDATORS = """
 [validator_list_sites]
@@ -49,96 +47,6 @@ https://vl.altnet.rippletest.net
 [validator_list_keys]
 ED264807102805220DA0F312E71FC2C69E1552C9C5790F6C25E3729DEB573D5860
 """
-
-MAINCHAIN_IP = "127.0.0.1"
-
-FEDERATORS_STANZA_INIT = """
-# federator signing public keys
-[sidechain_federators]
-"""
-FEDERATORS_SECRETS_STANZA_INIT = """
-# federator signing secret keys (for standalone-mode testing only; Normally won't be in
-# a config file)
-[sidechain_federators_secrets]
-"""
-BOOTSTRAP_FEDERATORS_STANZA_INIT = """
-# first value is federator signing public key, second is the signing pk account
-[sidechain_federators]
-"""
-
-
-def amt_to_json(amt):
-    if isinstance(amt, str):
-        return amt
-    else:
-        return amt.to_dict()
-
-
-def generate_asset_stanzas(assets: Optional[Dict[str, XChainAsset]] = None) -> str:
-    if assets is None:
-        # default to xrp only at a 1:1 value
-        assets = {}
-        assets["xrp_xrp_sidechain_asset"] = XChainAsset("0", "0", 1, 1, 400, 400)
-
-    index_stanza = """
-[sidechain_assets]"""
-
-    asset_stanzas = []
-
-    for name, xchainasset in assets.items():
-        index_stanza += "\n" + name
-        new_stanza = f"""
-[{name}]
-mainchain_asset={json.dumps(amt_to_json(xchainasset.main_asset))}
-sidechain_asset={json.dumps(amt_to_json(xchainasset.side_asset))}
-mainchain_refund_penalty={json.dumps(amt_to_json(xchainasset.main_refund_penalty))}
-sidechain_refund_penalty={json.dumps(amt_to_json(xchainasset.side_refund_penalty))}"""
-        asset_stanzas.append(new_stanza)
-
-    return index_stanza + "\n" + "\n".join(asset_stanzas)
-
-
-# First element of the returned tuple is the sidechain stanzas
-# second element is the bootstrap stanzas
-def generate_sidechain_stanza(
-    mainchain_ports: Ports,
-    main_account: Wallet,
-    federators: List[Keypair],
-    signing_key: str,
-    mainchain_cfg_file: str,
-    xchain_assets: Optional[Dict[str, XChainAsset]] = None,
-) -> Tuple[str, str]:
-    assets_stanzas = generate_asset_stanzas(xchain_assets)
-
-    federators_stanza = FEDERATORS_STANZA_INIT
-    federators_secrets_stanza = FEDERATORS_SECRETS_STANZA_INIT
-    bootstrap_federators_stanza = BOOTSTRAP_FEDERATORS_STANZA_INIT
-    for fed in federators:
-        federators_stanza += f"{fed.public_key}\n"
-        federators_secrets_stanza += f"{fed.secret_key}\n"
-        bootstrap_federators_stanza += f"{fed.public_key} {fed.account_id}\n"
-
-    sidechain_stanzas = f"""
-[sidechain]
-signing_key={signing_key}
-mainchain_account={main_account.classic_address}
-mainchain_ip={MAINCHAIN_IP}
-mainchain_port_ws={mainchain_ports.ws_public_port}
-# mainchain config file is: {mainchain_cfg_file}
-
-{assets_stanzas}
-
-{federators_stanza}
-
-{federators_secrets_stanza}
-"""
-    bootstrap_stanzas = f"""
-[sidechain]
-mainchain_secret={main_account.seed}
-
-{bootstrap_federators_stanza}
-"""
-    return (sidechain_stanzas, bootstrap_stanzas)
 
 
 # cfg_type will typically be either 'dog' or 'test', but can be any string. It is only
